refactor(chore): 按项目命名规则重命名文件
 - DefaultAnimation.ts -> defaultAnimation.ts
 - DefatultTheme.ts -> defaultTheme.ts
 - history/History.ts -> history/index.ts
 - 并更新项目中各个文件的引用地址
