+++ packages/core/src/LogicFlow.tsx
@@ -16,11 +16,11 @@ import Graph from './view/Graph'
 import * as _View from './view'
 import { formatData } from './util'
 
-import Dnd from './view/behavior/DnD'
+import Dnd from './view/behavior/dnd'
 import Tool from './tool/tool'
 import { snapline } from './tool'
 import Keyboard from './keyboard'
-import History from './history/History'
+import History from './history'
 import { CallbackType, EventArgs } from './event/eventEmitter'
 import { ElementType, EventType, SegmentDirection } from './constant'
 import { initDefaultShortcut } from './keyboard/shortcut'
