+++ packages/core/src/constant/defaultAnimation.ts
